Refactor XtdfDetectorBase._guess_axes() to be more efficient
Previously when `unstack_pulses` was set, the dimensions in train_pulse_ids were
unstacked using `DataArray.unstack()`. Unfortunately this silently made a copy,
but also inserted the new dimensions at the end (per the documentation) which is
not where they should go. The dimensions were reordered with
`DataArray.transpose()`, but since a copy was already made with the
train_pulse_ids dimensions in the wrong place, the transpose would return a view
of the array that was not in memory order. Now we figure out the right order of
dimensions ahead of time and use `.reshape()` to manually reshape the array
before creating a `DataArray`.

But not all copies are avoided. If the selection has a varying number of frames
per train then the frames are copied to the right location. In the future it
might be better to load them into the right location immediately.