Optimize slow path for KeyData.data_counts(labelled=False)
KeyData.data_counts() can return either a Series or ndarray, and for the ndarray
case there was a lot of manual work being done to make sure that the counts
array could handle missing trains and trains that are out-of-order, which was
relatively slow.

For a source with 85% of trains missing in a run with ~12,000 trains,
KeyData.data_counts(labelled=False) took 5.5ms. List of optimizations and
measurements after being applied:
- Vectorizing everything instead of building a dict: 2.3ms
- Using `assume_unique=True` in np.intersect1d(): 2.1ms
- Using `all_tids_arr` instead of `self.train_ids` (a list): 1.48ms
- Re-use `tid_to_ix` in the assert statement instead of using np.isin: 1.2ms

Decided not to pass `assume_unique=True`, so the final time was 1.45ms, about
~3.8x faster than the original version. Note that returning a Series is about
three orders of magnitude faster at about 122us.